Move startRebroadcastScheduler() and stopRebroadcastRescheduler() into WalletAppConfig
